Add event stream generator and route
